-Отделена логика взаим-я с почтой -Автозапуск таймер таски -Авторассылка -Автопереподключение -Удален крон -Рефакторинг кнопок
